Fix jni bugs
Former-commit-id: 930b7e5bcd16ee6abca7e4b2dfb9a181e33de5c2

+++ android/InspireFaceExample/inspireface/src/main/java/com/insightface/sdk/inspireface/InspireFace.java
@@ -254,7 +254,7 @@ public class InspireFace extends TypeDefine {
      * @param id Face identity
      * @return true if success, false otherwise
      */
-    public static native boolean FeatureHubFaceRemove(int id);
+    public static native boolean FeatureHubFaceRemove(long id);
 
     /**
      * Update face feature in feature hub
@@ -268,7 +268,7 @@ public class InspireFace extends TypeDefine {
      * @param id Face identity
      * @return FaceFeatureIdentity object
      */
-    public static native FaceFeatureIdentity FeatureHubGetFaceIdentity(int id);
+    public static native FaceFeatureIdentity FeatureHubGetFaceIdentity(long id);
 
     /**
      * Get face count in feature hub
